workers/constructors/SharedWorker/null-arguments.html

<!doctype html>
<title>null as arguments</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>
async_test(function() {
  const worker = new SharedWorker(null, null);
  worker.port.onmessage = this.step_func_done(function(e) {
    assert_equals(e.data[0], 'null', 'first arg');
    assert_equals(e.data[1], '', 'second arg');
  });
}, 'Test constructing a shared worker with null');
</script>
